+++ app/src/main/AndroidManifest.xml
@@ -19,7 +19,7 @@
         android:theme="@style/Theme.VRE_App"
         tools:targetApi="31">
         <activity
-            android:name=".KeyPhraseActivity"
+            android:name=".view.KeyPhraseActivity"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
@@ -28,7 +28,7 @@
 
         <activity
 
-            android:name=".ListenSpeechActivity"
+            android:name=".view.ListenSpeechActivity"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
@@ -36,21 +36,21 @@
         </activity>
         <activity
 
-            android:name=".ForgotPasswordActivity"
+            android:name=".view.ForgotPasswordActivity"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".MenuActivity"
+            android:name=".view.MenuActivity"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".LoginActivity"
+            android:name=".view.LoginActivity"
 
             android:exported="true">
             <intent-filter>
@@ -64,14 +64,14 @@
                 android:value="" />
         </activity>
         <activity
-            android:name=".CreateAccountActivity"
+            android:name=".view.CreateAccountActivity"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".MainActivity"
+            android:name=".view.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
