+++ app/src/main/java/com/metrostateics499/vre_app/model/data/KeyPhrase.kt
@@ -1,4 +1,4 @@
-package com.metrostateics499.vre_app
+package com.metrostateics499.vre_app.model.data
 
 /**
  * Key phrase
