+++ app/src/main/java/com/metrostateics499/vre_app/model/data/KeyPhraseList.kt
@@ -1,4 +1,6 @@
-package com.metrostateics499.vre_app
+package com.metrostateics499.vre_app.model.data
+
+import com.metrostateics499.vre_app.model.data.KeyPhrase
 
 data class KeyPhraseList(var keyPhrases: MutableList<KeyPhrase>) {
