+++ app/src/main/java/com/metrostateics499/vre_app/view/CreateAccountActivity.kt
@@ -1,4 +1,4 @@
-package com.metrostateics499.vre_app
+package com.metrostateics499.vre_app.view
 
 import android.content.Context
 import android.content.Intent
@@ -8,6 +8,7 @@ import android.widget.Button
 import android.widget.EditText
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
+import com.metrostateics499.vre_app.R
 import com.metrostateics499.vre_app.model.Passing
 
 class CreateAccountActivity : AppCompatActivity() {
@@ -37,16 +38,16 @@ class CreateAccountActivity : AppCompatActivity() {
 
     private fun validateEmptyForm() {
         if (fullName.text.toString().isNotEmpty() && userName.text.toString()
-            .isNotEmpty() && email.text.toString().isNotEmpty() && password.text.toString()
-                .isNotEmpty() && confirmPassword.text.toString().isNotEmpty()
+                    .isNotEmpty() && email.text.toString().isNotEmpty() && password.text.toString()
+                    .isNotEmpty() && confirmPassword.text.toString().isNotEmpty()
         ) {
             // Need a valid Email to work (ex. abc123@gmail.net)
             if (email.length() in 6..319 && email.text.toString()
-                .matches(Regex("[a-zA-Z0-9._]+@[a-z].+[a-z]"))
+                        .matches(Regex("[a-zA-Z0-9._]+@[a-z].+[a-z]"))
             ) {
                 // Maximum length for the full name, username, password, email and confirm password
                 if (fullName.length() in 6..36 && userName.length() in 4..36 &&
-                    password.length() in 8..36 && confirmPassword.length() in 8..36
+                        password.length() in 8..36 && confirmPassword.length() in 8..36
                 ) {
                     if (password.text.toString() == confirmPassword.text.toString()) {
                         Toast.makeText(
