+++ app/src/main/java/com/metrostateics499/vre_app/view/ForgotPasswordActivity.kt
@@ -1,7 +1,8 @@
-package com.metrostateics499.vre_app
+package com.metrostateics499.vre_app.view
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import com.metrostateics499.vre_app.R
 
 class ForgotPasswordActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
