+++ app/src/main/java/com/metrostateics499/vre_app/view/LoginActivity.kt
@@ -1,4 +1,4 @@
-package com.metrostateics499.vre_app
+package com.metrostateics499.vre_app.view
 
 import android.content.Intent
 import android.os.Bundle
@@ -7,6 +7,7 @@ import android.widget.EditText
 import android.widget.TextView
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
+import com.metrostateics499.vre_app.R
 import com.metrostateics499.vre_app.model.Passing
 
 class LoginActivity : AppCompatActivity() {
@@ -59,7 +60,7 @@ class LoginActivity : AppCompatActivity() {
 
         // Cannot able to login with an empty text
         if (username.text.toString().isEmpty() && password.text.toString()
-            .isEmpty() && email.text.isEmpty()
+                    .isEmpty() && email.text.isEmpty()
         ) {
             Toast.makeText(
                 this,
@@ -67,10 +68,10 @@ class LoginActivity : AppCompatActivity() {
                 Toast.LENGTH_SHORT
             ).show()
         } else if ((
-            (inputUsername == Passing.username && inputPassword == Passing.password) ||
-                (inputUsername == "username" && inputPassword == "password") ||
-                (inputEmail == Passing.email && inputPassword == Passing.password)
-            )
+                        (inputUsername == Passing.username && inputPassword == Passing.password) ||
+                                (inputUsername == "username" && inputPassword == "password") ||
+                                (inputEmail == Passing.email && inputPassword == Passing.password)
+                        )
         ) {
 
             Toast.makeText(this, "Successfully Logged In", Toast.LENGTH_LONG).show()
