+++ app/src/main/res/layout/activity_create_account.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#FF185CA5"
-    tools:context=".CreateAccountActivity">
+    tools:context=".view.CreateAccountActivity">
 
     <EditText
         android:id="@+id/editTextPersonName"
