+++ app/src/main/res/layout/activity_forgot_password.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#FF185CA5"
-    tools:context=".ForgotPasswordActivity">
+    tools:context=".view.ForgotPasswordActivity">
 
     <TextView
         android:id="@+id/app_logo"
